# baremaps-ogcapi/src/main/resources/studio-openapi.yaml
openapi: 3.0.3
info:
  title: Baremaps API
  description: Draft for a Baremaps API.
  version: ${project.version}
servers:
  - url: 'http://localhost:8080/studio'
...
paths:
...
  /maps:
    get:
      tags:
        - Maps
      summary: the maps metadata
      description: A list of all maps metadata available.
      operationId: getMaps
      responses:
        "200":
          description: |-
            The maps shared by this API.
            This response can be references directly for every service that wants only essential information at the maps level. /maps/mapId might return more information.
            This resource provides information about and how to access the maps.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/maps-metadata'
        "500":
          description: A server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
    post:
      tags:
        - Maps
      summary: |-
        adds a new map
      operationId: addMap
      description: |-
        Adds a map to the map repository.

        In case, a new map is created, the following rules
        apply.

        If the map submitted in the request body includes an
        identifier, that identifier will be used. If a map with that identifier
        already exists, an error is returned.

        If no identifier can be determined from the submitted
        map, the server will assign a new identifier to the
        map.

        The URI of the new map is returned in the header
        `Location`.
      responses:
        '201':
          description: |-
            map created
          headers:
            Location:
              schema:
                type: string
              description: |-
                URI of the new map
        '409':
          description: |-
            a map with that id already exists
      requestBody:
        description: |-
          Map to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/map-metadata'
...
components:
...
  schemas:
...
    exception:
      type: object
      required:
        - code
      properties:
        code:
          type: string
        description:
          type: string
...
    map-metadata:
      type: object
      required:
        - id
      properties:
        id:
          description: identifier of the map used, for example, in URIs
          type: string
          format: uuid
        title:
          description: human readable title of the map
          type:
            string
        style:
          description: id/reference to the style used in the map
          type: string
        tileset:
          description: id/reference to the tileset used in the map
          type: string
        last-edit:
          description: date of the last edit of the map
          type: string
        publication-scope:
          description: indication if its a private or published map
          type: string
        views:
          description: count of map views
          type: integer
    maps-metadata:
      type: object
      properties:
        maps-metadata:
          type: array
          items:
            $ref: '#components/schemas/map-metadata'
...